* Add sample config for future self hosters

* WIP: Initial commit

* Fix !eval code block error

* please just don't just to just just, just

* Linker progress

* Quarantine dev guilds

* Fix custom wiki param commands

* Too much for one commit message

* Remove legacy linker code after refactor

* Dev quarantine quoter

* Linker upgrades

* Don't do anything if >role commands have content

* Clean up deps

+++ src/plugins/commander/commands/lua.js
@@ -14,7 +14,9 @@ class LuaRoleCommand extends Command {
         ];
     }
 
-    call(message) {
+    call(message, content) {
+        if (content) return;
+
         message.delete();
         if (message.member.roles.has('269869890087682049')) {
             message.member.removeRole('269869890087682049');
